ml-agents/requirements.txt
# Core ML/AI
openai-whisper==20231117
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# NLP
spacy==3.7.2
nltk==3.8.1

# FastAPI for agent API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
motor==3.3.2
pymongo==4.6.0

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

